+++ src/aggregate.h
@@ -132,7 +132,6 @@ public:
     virtual Scope *newScope(Scope *sc);
     void setScope(Scope *sc);
     void semantic2(Scope *sc);
-    void semantic3(Scope *sc);
     bool determineFields();
     bool determineSize(Loc loc);
     virtual void finalizeSize() = 0;
