+++ src/compiler.c
@@ -149,7 +149,7 @@ void Compiler::genCmain(Scope *sc)
     m->importAll(NULL);
     m->semantic(NULL);
     m->semantic2(NULL);
-    m->semantic3(NULL);
+    semantic3(m, NULL);
     global.params.verbose = v;
 
     entrypoint = m;
