+++ src/dclass.c
@@ -878,7 +878,7 @@ Lancestorsdone:
     if (deferred && !global.gag)
     {
         deferred->semantic2(sc);
-        deferred->semantic3(sc);
+        semantic3(deferred, sc);
     }
     //printf("-ClassDeclaration::semantic(%s), type = %p, sizeok = %d, this = %p\n", toChars(), type, sizeok, this);
 }
